Conclusão do recyclerView, porém exibe a recyclerView vazia
